add rules related to css-tokens
