+++ packages/eslint-plugin-vapor/src/utils/color-resolver.ts
@@ -1,4 +1,4 @@
-import { formatHex, formatHex8, useMode, modeOklch } from 'culori/fn';
+import { formatHex, formatHex8, modeOklch, useMode } from 'culori/fn';
 
 // Register oklch mode so formatHex / formatHex8 can convert oklch → rgb
 useMode(modeOklch);
@@ -6,13 +6,13 @@ useMode(modeOklch);
 type OklchColor = { mode: 'oklch'; l: number; c: number; h: number; alpha?: number };
 
 export function oklchToHex(components: readonly [number, number, number], alpha?: number): string {
-  const color: OklchColor = {
-    mode: 'oklch',
-    l: components[0],
-    c: components[1],
-    h: components[2],
-    alpha: alpha,
-  };
-  const result = alpha != null && alpha < 1 ? formatHex8(color) : formatHex(color);
-  return (result ?? '#000000').toLowerCase();
+    const color: OklchColor = {
+        mode: 'oklch',
+        l: components[0],
+        c: components[1],
+        h: components[2],
+        alpha: alpha,
+    };
+    const result = alpha != null && alpha < 1 ? formatHex8(color) : formatHex(color);
+    return (result ?? '#000000').toLowerCase();
 }
