+++ packages/eslint-plugin-vapor/src/utils/token-scope.ts
@@ -1,21 +1,21 @@
 import type { Scope } from '~/data/property-scope-map';
 
 export function scopeFromTokenName(name: string): Scope | 'primitive' | null {
-  if (!name.startsWith('--vapor-')) return null;
-  const segments = name.slice('--vapor-'.length).split('-');
-  if (segments[0] === 'color') {
-    const second = segments[1];
-    if (second === 'foreground' || second === 'background' || second === 'border') {
-      return second;
+    if (!name.startsWith('--vapor-')) return null;
+    const segments = name.slice('--vapor-'.length).split('-');
+    if (segments[0] === 'color') {
+        const second = segments[1];
+        if (second === 'foreground' || second === 'background' || second === 'border') {
+            return second;
+        }
+        return 'primitive';
     }
-    return 'primitive';
-  }
-  if (segments[0] === 'size') {
-    const second = segments[1];
-    if (second === 'dimension' || second === 'space' || second === 'borderRadius') {
-      return second;
+    if (segments[0] === 'size') {
+        const second = segments[1];
+        if (second === 'dimension' || second === 'space' || second === 'borderRadius') {
+            return second;
+        }
     }
-  }
-  if (segments[0] === 'shadow') return 'shadow';
-  return null;
+    if (segments[0] === 'shadow') return 'shadow';
+    return null;
 }
